Phase-III-AI-Powered-Todo-App/backend/alembic/versions/8730e2bda261_initial_schema_setup.py
"""Initial schema setup

Revision ID: 8730e2bda261
Revises: 
Create Date: 2026-02-08 15:17:51.838823

"""
from typing import Sequence, Union

from alembic import op
import sqlalchemy as sa


# revision identifiers, used by Alembic.
revision: str = '8730e2bda261'
down_revision: Union[str, Sequence[str], None] = None
branch_labels: Union[str, Sequence[str], None] = None
depends_on: Union[str, Sequence[str], None] = None


def upgrade() -> None:
    """Upgrade schema."""
    pass


def downgrade() -> None:
    """Downgrade schema."""
    pass
