Phase-III-AI-Powered-Todo-App/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0.post1"}
sqlmodel = "0.0.14"
psycopg2-binary = "^2.9.9" # For PostgreSQL
alembic = "^1.13.1" # For migrations
pyjwt = "^2.11.0"
httpx = "^0.28.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
ruff = "^0.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
